Fix build warnings in mhi and qrtr.

+++ drivers/bus/mhi/core/mhi_main.c
@@ -1904,8 +1904,8 @@ int mhi_debugfs_mhi_event_show(struct seq_file *m, void *d)
 			seq_printf(m,
 				   " rp:0x%llx wp:0x%llx local_rp:0x%llx db:0x%llx\n",
 				   er_ctxt->rp, er_ctxt->wp,
-				   mhi_to_physical(ring, ring->rp),
-				   mhi_event->db_cfg.db_val);
+				   (unsigned long long)mhi_to_physical(ring, ring->rp),
+				   (unsigned long long)mhi_event->db_cfg.db_val);
 		}
 	}
 
@@ -1938,9 +1938,9 @@ int mhi_debugfs_mhi_chan_show(struct seq_file *m, void *d)
 				   " base:0x%llx len:0x%llx wp:0x%llx local_rp:0x%llx local_wp:0x%llx db:0x%llx\n",
 				   chan_ctxt->rbase, chan_ctxt->rlen,
 				   chan_ctxt->wp,
-				   mhi_to_physical(ring, ring->rp),
-				   mhi_to_physical(ring, ring->wp),
-				   mhi_chan->db_cfg.db_val);
+				   (unsigned long long)mhi_to_physical(ring, ring->rp),
+				   (unsigned long long)mhi_to_physical(ring, ring->wp),
+				   (unsigned long long)mhi_chan->db_cfg.db_val);
 		}
 	}
